Move utility functions to utils.py
